efs_project/requirements.txt
Django>=5.2.1
Pillow>=11.2.1  # для работы с изображениями
python-dotenv>=1.1.0  # для работы с переменными окружения
django-allauth>=65.8.0  # для аутентификации
django-crispy-forms>=2.0  # для форм
crispy-bootstrap5>=0.7  # для стилизации форм
django-nested-admin>=4.1.1  # для вложенных форм в админке
django-widget-tweaks>=1.5.0  # для настройки виджетов форм
requests>=2.32.3  # для HTTP-запросов
PyJWT>=1.3.1  # для работы с JWT токенами
asgiref>=3.8.1  # ASGI спецификация
certifi>=2025.4.26  # сертификаты для HTTPS
cffi>=1.17.1  # Foreign Function Interface
charset-normalizer>=3.4.2  # нормализация кодировок
cryptography>=44.0.3  # криптографические функции
idna>=3.10  # поддержка IDNA
pycparser>=2.22  # парсер C
sqlparse>=0.5.3  # парсер SQL
tzdata>=2025.2  # база данных часовых поясов
urllib3>=2.4.0  # HTTP клиент
psycopg2-binary>=2.9.9  # PostgreSQL адаптер для Python 